+++ include/ccf/service/node_info_network.h
@@ -5,6 +5,7 @@
 
 #include "ccf/ds/json.h"
 #include "ccf/ds/nonstd.h"
+#include "ccf/http_configuration.h"
 #include "ccf/service/acme_client_config.h"
 
 #include <string>
@@ -56,22 +57,36 @@ namespace ccf
 
   static constexpr auto PRIMARY_RPC_INTERFACE = "ccf.default_rpc_interface";
 
+  /// Node network information
   struct NodeInfoNetwork_v2
   {
     using NetAddress = std::string;
     using RpcInterfaceID = std::string;
     using NetProtocol = std::string;
 
+    /// Network interface description
     struct NetInterface
     {
       NetAddress bind_address;
       NetAddress published_address;
       NetProtocol protocol;
 
+      /// Maximum open sessions soft limit
       std::optional<size_t> max_open_sessions_soft = std::nullopt;
+
+      /// Maximum open sessions hard limit
       std::optional<size_t> max_open_sessions_hard = std::nullopt;
 
+      /// HTTP configuration
+      std::optional<http::ParserConfiguration> http_configuration =
+        std::nullopt;
+
+      /// Interface endorsement
       std::optional<Endorsement> endorsement = std::nullopt;
+
+      /// Regular expressions of endpoints that are accessible over
+      /// this interface. std::nullopt means everything is accepted.
+      std::optional<std::vector<std::string>> accepted_endpoints = std::nullopt;
 
       bool operator==(const NetInterface& other) const
       {
@@ -80,24 +95,34 @@ namespace ccf
           protocol == other.protocol &&
           max_open_sessions_soft == other.max_open_sessions_soft &&
           max_open_sessions_hard == other.max_open_sessions_hard &&
-          endorsement == other.endorsement;
+          endorsement == other.endorsement &&
+          http_configuration == other.http_configuration &&
+          accepted_endpoints == other.accepted_endpoints;
       }
     };
 
+    /// RPC interface mapping
     using RpcInterfaces = std::map<RpcInterfaceID, NetInterface>;
 
+    /// Node-to-node network interface
     NetInterface node_to_node_interface;
+
+    /// RPC interfaces
     RpcInterfaces rpc_interfaces;
 
+    /// ACME configuration description
     struct ACME
     {
+      /// Mapping of ACME client configuration names to configurations
       std::map<std::string, ccf::ACMEClientConfig> configurations;
 
       bool operator==(const ACME&) const = default;
     };
 
-    std::optional<ACME> acme;
+    /// ACME configuration
+    std::optional<ACME> acme = std::nullopt;
   };
+
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(NodeInfoNetwork_v2::NetInterface);
   DECLARE_JSON_REQUIRED_FIELDS(NodeInfoNetwork_v2::NetInterface, bind_address);
   DECLARE_JSON_OPTIONAL_FIELDS(
@@ -106,7 +131,9 @@ namespace ccf
     max_open_sessions_soft,
     max_open_sessions_hard,
     published_address,
-    protocol);
+    protocol,
+    http_configuration,
+    accepted_endpoints);
   DECLARE_JSON_TYPE(NodeInfoNetwork_v2::ACME);
   DECLARE_JSON_REQUIRED_FIELDS(NodeInfoNetwork_v2::ACME, configurations);
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(NodeInfoNetwork_v2);
@@ -191,8 +218,7 @@ namespace ccf
   }
 }
 
-FMT_BEGIN_NAMESPACE
-template <>
+FMT_BEGIN_NAMESPACE template <>
 struct formatter<ccf::Authority>
 {
   template <typename ParseContext>