+++ CHANGELOG.md
@@ -10,12 +10,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `/node/version` now contains an `unsafe` flag reflecting the status of the build.
+- New per-interface configuration entries (`network.rpc_interfaces.http_configuration`) are added to let operators cap the maximum size of body, header value size and number of headers in client HTTP requests. The client session is automatically closed if the HTTP request exceeds one of these limits (#3941).
+- Added new `recovery_count` field to `GET /node/network` endpoint to track the number of disaster recovery procedures undergone by the service (#3982).
 
 ### Changed
 
 - Generated OpenAPI now describes whether each endpoint is forwarded (#3935).
 
+### Fixed
+
+- Fixed issue with recovery of large ledger entries (#3986).
+
 ### Documentation
+
 - The "Node Output" page has been relabelled as "Troubleshooting" in the documentation and CLI commands for troubleshooting have been added to it.
 
 ## [3.0.0-dev0]